Move App.tsx to src folder, update image names
