+++ client/src/components/UsersContainer.tsx
@@ -8,21 +8,20 @@ import Users from './Users';
 import {IUserContainerProps} from '../interfaces/User.interface'
 
 export default function UsersContainer({ users }: IUserContainerProps) {
-
   const { currentUser } = useAuth();
   return (
     <Row
       xs={1}
-      md={2}
+      md={4}
       id='users-container'
       className='g-4 mt-auto'
     >
       {users.map(user => (
         user.uid === currentUser.uid
           ? null
-          : (<Col key={user.name}>
+          : <Col key={user.uid}>
             <Users user={user} />
-          </Col>)
+          </Col>
       ))}
     </Row>
   );
